doc/demo: Drop /sbin from dhclient command, pass script file explicitly
dhclient might be in /usr/sbin on recent versions of Fedora, and
mbuto just adds it to the same location as it was on the host:
just call dhclient instead of /sbin/dhclient.

This also applies for dhclient-script: given that we create the file
on boot, pass its explicit location with -sf.

Signed-off-by: Stefano Brivio <sbrivio@redhat.com>